bot-server/requirements.txt
aiohttp==3.9.5
supabase==2.5.0
cryptography==42.0.5
python-dotenv==1.0.1
